Completely changed download, delete, download systems.
