--- LocalFiles/command_auth.txt
@@ -1,1 +1,0 @@
-%OK
