--- ServerFiles/cmd/command_download.txt
@@ -1,1 +1,0 @@
-Serverfile2.txt
