--- ServerFiles/command_delete.txt
@@ -1,1 +1,0 @@
-todelete.txt
