+++ ServerFiles/java_cloud_gb_hw5.txt
@@ -1,0 +1,22 @@
+Реализовал возможность скачивания, удаления и аутентификации через своего рода костыль: название файла записывается в заранее созданный 
+файл и отправляется в формате AbstractMessage. На стороне fileHandler обрабатывается получение файла.
+
+За неделю:
+
+1. Реализовал авторизацию через базу данных. Когда в проекте всего один пользователь, в этом нет необходимости. Но в перспевтиве есть возможность 
+сделать для каждого пользователя свою "папку на серверной стороне". В таком случае авторизация будет полезна.
+
+2. Немного изменил интерфейс
+
+
+Предстоит:
+
+1. Обработать случай неверного введения логина и пароля в jfx в виде предупреждения
+
+2. Реализовать возможность перемеинования файла на сервере;
+
+3. Упростить и оптимизировать код, связанный с отправкой и обработкой файлов;
+
+4. Сделать выпадающее меню в jfx: при нажатии правой кнопкой мыши на файл, выпадает менюшка с выбором, что делать с файлом вместо кнопок;
+
+5. Сделать интерфейс более привлекательным и удобным;
