+++ gb-cloud-application/src/main/java/ru/gb/cloud/controllers/MainController.java
@@ -9,19 +9,13 @@ import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
 import lombok.extern.slf4j.Slf4j;
-import ru.gb.cloud.model.AbstractMessage;
-import ru.gb.cloud.model.DownloadMessage;
-import ru.gb.cloud.model.FileMessage;
-import ru.gb.cloud.model.ListMessage;
+import ru.gb.cloud.model.*;
 import ru.gb.cloud.network.Net;
 
-import java.io.BufferedWriter;;
 import java.io.IOException;
 import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -51,48 +45,35 @@ public class MainController implements Initializable {
 
     @FXML HBox loginBox;
 
-
-   // folder where the "download_command.txt" is
-   private final Path cmdFiles = Path.of("cmd files");
-
-   // exact path of "download_command.txt"
-   private final Path fileDownDir = Path.of("cmd files/command_download.txt");
-
-   private final Path fileDelete = Path.of("cmd files/command_delete.txt");
-
-   private final Path authDataDir = Path.of("cmd files/command_auth.txt");
-
     private void read() {
         try {
             while (true) {
                 AbstractMessage message = net.read();
+
                 if (message instanceof ListMessage lm) {
                     serverView.getItems().clear();
                     serverView.getItems().addAll(lm.getFiles());
                 }
-                if (message instanceof FileMessage file) {
-                    // if the message is FileMessage, download the file
-                    Files.write(clientDir.resolve(file.getName()), file.getBytes());
-                    // decode the file and get the status
-                    byte[] encodedAuth = Files.readAllBytes(Paths.get("LocalFiles/command_auth.txt"));
-                    String status = new String(encodedAuth, StandardCharsets.UTF_8);
 
-                    // if status OK make all visible
+                if (message instanceof FileMessage file) {
+                    Files.write(clientDir.resolve(file.getName()), file.getBytes());
+                    reloadList();
+                }
+
+                if (message instanceof AuthMessage authMessage) {
+                    String status = authMessage.getAuthData();
                     if (status.equals("%OK")) {
-                       loginBox.setVisible(false);
-                       clientView.setVisible(true);
-                       serverView.setVisible(true);
-                       deleteButton.setVisible(true);
-                       uploadButton.setVisible(true);
-                       downloadButton.setVisible(true);
+                        loginBox.setVisible(false);
+                        clientView.setVisible(true);
+                        serverView.setVisible(true);
+                        deleteButton.setVisible(true);
+                        uploadButton.setVisible(true);
+                        downloadButton.setVisible(true);
                     }
 
-                    if (message.equals("%WrongData")){
+                    if (message.equals("%WrongData")) {
                         System.out.println("Wrong login or password");
                     }
-                    
-                    clientView.getItems().clear();
-                    clientView.getItems().addAll(getClientFiles());
                 }
 
             }
@@ -125,17 +106,7 @@ public class MainController implements Initializable {
     }
 
     public void download(ActionEvent actionEvent) throws Exception {
-        // choosing a fileName with a click in jfx
         String downloadFile = serverView.getSelectionModel().getSelectedItem();
-        System.out.println("selectedFileDownload: " + downloadFile);
-        // writes the name of the file to be downloaded into the "command_download.txt"
-       // Files.writeString(fileDownDir, downloadFile, StandardCharsets.UTF_8);
-        // sends the "command_download.txt" as a FileMessage
-       // net.write(new FileMessage(cmdFiles.resolve("command_download.txt")));
-        // clears the "command_download.txt" file for a further usage
-        //BufferedWriter writer = Files.newBufferedWriter(Paths.get(String.valueOf(fileDownDir)));
-        //writer.write("");
-       // writer.flush();
         net.write(new DownloadMessage(downloadFile));
     }
 
@@ -145,16 +116,8 @@ public class MainController implements Initializable {
     }
 
     public void delete(ActionEvent actionEvent) throws IOException {
-        // get name of the folder to be deleted
         String fileToDelete = serverView.getSelectionModel().getSelectedItem();
-        // write the name into the file
-        Files.writeString(fileDelete, fileToDelete, StandardCharsets.UTF_8);
-        // send the file
-        net.write(new FileMessage(cmdFiles.resolve("command_delete.txt")));
-        // clear the local file
-        BufferedWriter writer = Files.newBufferedWriter(Paths.get(String.valueOf(fileDelete)));
-        writer.write("");
-        writer.flush();
+        net.write(new DeleteMessage(fileToDelete));
     }
 
     public void btnAuthClick(ActionEvent actionEvent) throws IOException {
@@ -162,13 +125,6 @@ public class MainController implements Initializable {
         String password = passwordField.getText();
         String authData = login + "#" + password;
         System.out.println(authData);
-        // write data into the file
-        Files.writeString(authDataDir, authData, StandardCharsets.UTF_8);
-        // send the file
-        net.write(new FileMessage(cmdFiles.resolve("command_auth.txt")));
-        // clear the file
-        BufferedWriter writer = Files.newBufferedWriter(Paths.get(String.valueOf(authData)));
-        writer.write("");
-        writer.flush();
+        net.write(new AuthMessage(authData));
     }
 }
