+++ gb-cloud-core/src/main/java/ru/gb/cloud/model/AuthMessage.java
@@ -1,0 +1,19 @@
+package ru.gb.cloud.model;
+
+public class AuthMessage extends AbstractMessage {
+
+    public String getAuthData() {
+        return authData;
+    }
+
+    private final String authData;
+
+    public AuthMessage(String data) {
+        this.authData = data;
+    }
+
+    @Override
+    public MessageType getMessageType() {
+        return MessageType.AUTH;
+    }
+}
