+++ gb-cloud-core/src/main/java/ru/gb/cloud/model/DeleteMessage.java
@@ -1,0 +1,22 @@
+package ru.gb.cloud.model;
+
+import lombok.Getter;
+
+@Getter
+public class DeleteMessage extends AbstractMessage {
+
+    public String getDeleteFileName() {
+        return deleteFileName;
+    }
+
+    private final String deleteFileName;
+
+    public DeleteMessage(String deleteFile) {
+        this.deleteFileName = deleteFile;
+    }
+
+    @Override
+    public MessageType getMessageType() {
+        return MessageType.DELETE;
+    }
+}
