+++ gb-cloud-server/src/main/java/ru/gb/cloud/netty/serialization/SerializationPipeLine.java
@@ -2,12 +2,9 @@ package ru.gb.cloud.netty.serialization;
 
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.socket.SocketChannel;
-import io.netty.handler.codec.serialization.ClassResolver;
 import io.netty.handler.codec.serialization.ClassResolvers;
 import io.netty.handler.codec.serialization.ObjectDecoder;
 import io.netty.handler.codec.serialization.ObjectEncoder;
-import io.netty.handler.codec.string.StringDecoder;
-import io.netty.handler.codec.string.StringEncoder;
 import ru.gb.cloud.controllers.LocalFileHandler;
 
 public class SerializationPipeLine extends ChannelInitializer<SocketChannel> {
@@ -16,7 +13,7 @@ public class SerializationPipeLine extends ChannelInitializer<SocketChannel> {
         socketChannel.pipeline().addLast(
                 new ObjectDecoder(ClassResolvers.cacheDisabled(null)),
                 new ObjectEncoder(),
-                new FileHandler(),
+                new FileAndAuthHandler(),
                 new LocalFileHandler()
 
         );
